Update profile button in navigation bars
