+++ frontend/src/components/AdminNavBar.tsx
@@ -1,8 +1,10 @@
 import React from "react";
 import { useState } from "react";
 import { useLocation, Link } from "react-router-dom";
-import ProfileButton from "./ProfileModals/ProfileButton.tsx";
+import ProfileButton from "./ProfileButton.tsx";
 import AddQuestionModal from "./QuestionModals/AddQuestionModal.tsx";
+import { User } from "../types/User";
+import useRetrieveUser from "../hooks/useRetrieveUser";
 
 interface AdminNavBarProps {
   fetchData: () => Promise<void>;
@@ -14,7 +16,8 @@ const AdminNavBar: React.FC<AdminNavBarProps> = ({ fetchData }) => {
   const [isAddModalOpen, setAddModalOpen] = useState(false);
   const openAddModal = () => setAddModalOpen(true);
   const closeAddModal = () => setAddModalOpen(false);
-
+  const [user, setUser] = useState<User | undefined>(undefined);
+  useRetrieveUser(setUser);
   return (
     <nav className="bg-off-white w-full p-4 flex items-center justify-between">
       {/* Logo or Brand */}
@@ -41,9 +44,9 @@ const AdminNavBar: React.FC<AdminNavBarProps> = ({ fetchData }) => {
           </div>
         ))}
       {/* Profile button */}
-      <div>
+      <div className="flex-none">
         <Link to="/profile">
-          <ProfileButton />
+          <ProfileButton userId={user ? user.id : ""}/>
         </Link>
       </div>
     </nav>
