+++ frontend/src/components/NavBar.tsx
@@ -1,11 +1,14 @@
-import React from "react";
+import React, { useState } from "react";
 import { useLocation, Link } from "react-router-dom";
 import IsConnected from "./IsConnected";
-import ProfileButton from "./ProfileModals/ProfileButton";
+import ProfileButton from "./ProfileButton";
+import { User } from "../types/User";
+import useRetrieveUser from "../hooks/useRetrieveUser";
 
 const NavBar: React.FC = () => {
   const location = useLocation();
-
+  const [user, setUser] = useState<User | undefined>(undefined);
+  useRetrieveUser(setUser);
   return (
     <nav className="bg-off-white w-full p-4 flex items-center justify-between relative">
       {/* Logo or Brand (Left-aligned) */}
@@ -39,7 +42,7 @@ const NavBar: React.FC = () => {
           </div>
         ) : (
           <Link to="/profile">
-            <ProfileButton />
+            <ProfileButton userId={user ? user.id : ""}/>
           </Link>
         )}
       </div>
