--- frontend/src/components/ProfileModals/ProfileButton.tsx
@@ -1,13 +1,0 @@
-import React from "react";
-
-const ProfileButton: React.FC = () => {
-  return (
-    <>
-      <div className="container px-2">
-        <button className="bg-black rounded-full p-6"></button>
-      </div>
-    </>
-  );
-};
-
-export default ProfileButton;
