+++ frontend/src/config/config.ts
@@ -3,6 +3,7 @@ const apiConfig = {
     userServiceUserUrl: "http://localhost:3001/users",
     userServiceAuthUrl: "http://localhost:3001/auth",
     profilePictureServiceUserUrl: "http://localhost:8081/users",
+    token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjY3MDEwNDZhNWUwZGFhODQzNmEyMjRlNSIsImlhdCI6MTcyODgyMzc5NSwiZXhwIjoxNzI4OTEwMTk1fQ.nN4ggV2RGqqIBSN4u4e8jPd4cJsBtU-1p97l9dboRb4"
 };
 
 export default apiConfig;
